refactor(helm): add data structure for local chart cache
The existing Helm libraries can only lint charts that are already expanded in a directory. In order to allow linting a tarball or a remote chart, it is required to expand it to a directory prior to it.

This change introduces a chart cache, where the in memory chart data-structure and the path it has been expanded for further inspection.